working on client

+++ api_wrapper/thecatapi/java/src/main/java/model/Image.java
@@ -6,8 +6,8 @@ import java.util.List;
 public class Image extends Model {
     private final String url;
     private final String mimeType;
-    private final int width;
-    private final int height;
+    private final double width;
+    private final double height;
     private final List<Breed> breeds;
     private final List<Category> categories;
 
@@ -15,8 +15,8 @@ public class Image extends Model {
         super(id, data);
         this.url = String.valueOf(data.get("url"));
         this.mimeType = String.valueOf(data.get("mime_type"));
-        this.width = (int) data.get("width");
-        this.height = (int) data.get("height");
+        this.width = (double) data.get("width");
+        this.height = (double) data.get("height");
         this.breeds = breeds;
         this.categories = categories;
     }
@@ -29,11 +29,11 @@ public class Image extends Model {
         return mimeType;
     }
 
-    public int getWidth() {
+    public double getWidth() {
         return width;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
